Backend/agent/requirements.txt
groq>=0.9.0
ddgs>=1.0.0
playwright>=1.44.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
